+++ CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v1.0.0-beta.1](https://github.com/ollm/OpenComic/releases/tag/v1.0.0-beta.1) (dd-mm-yyyy)
+## v1.0.0-beta.2 (dd-mm-yyyy)
+
+### 🚀 New Features
+
+- Reload button in file list
+
+### 🐛 Bug Fixes
+
+
+## [v1.0.0-beta.1](https://github.com/ollm/OpenComic/releases/tag/v1.0.0-beta.1) (22-08-2023)
 
 ### 🚀 New Features